requirements.txt
pynput==1.7.4
Pillow==8.4.0
unidecode==1.3.2
psutil==5.8.0
pywin32==302
